+++ src/main/scala/com/jcdecaux/setl/internal/HasReader.scala
@@ -1,0 +1,9 @@
+package com.jcdecaux.setl.internal
+
+import org.apache.spark.sql.DataFrameReader
+
+trait HasReader {
+
+  protected val reader: DataFrameReader
+
+}