+++ src/main/scala/com/jcdecaux/setl/storage/connector/CassandraConnector.scala
@@ -6,6 +6,7 @@ import com.datastax.spark.connector.cql.{CassandraConnector => DatastaxCassandra
 import com.jcdecaux.setl.annotation.InterfaceStability
 import com.jcdecaux.setl.config.Conf
 import com.jcdecaux.setl.enums.Storage
+import com.jcdecaux.setl.internal.HasReaderWriter
 import com.jcdecaux.setl.util.TypesafeConfigUtils
 import com.typesafe.config.Config
 import org.apache.spark.sql._
@@ -18,13 +19,7 @@ import org.apache.spark.sql.cassandra._
 class CassandraConnector(val keyspace: String,
                          val table: String,
                          val partitionKeyColumns: Option[Seq[String]],
-                         val clusteringKeyColumns: Option[Seq[String]]) extends DBConnector {
-
-  def this(keyspace: String,
-           table: String,
-           spark: SparkSession,
-           partitionKeyColumns: Option[Seq[String]],
-           clusteringKeyColumns: Option[Seq[String]]) = this(keyspace, table, partitionKeyColumns, clusteringKeyColumns)
+                         val clusteringKeyColumns: Option[Seq[String]]) extends DBConnector with HasReaderWriter {
 
   /**
    * Constructor with a [[com.jcdecaux.setl.config.Conf]] object
@@ -60,22 +55,6 @@ class CassandraConnector(val keyspace: String,
         None
       }
   )
-
-  /**
-   * Constructor with a [[com.jcdecaux.setl.config.Conf]] object
-   *
-   * @param spark spark session
-   * @param conf  [[com.jcdecaux.setl.config.Conf]] object
-   */
-  def this(spark: SparkSession, conf: Conf) = this(conf)
-
-  /**
-   * Constructor with a [[com.typesafe.config.Config]] object
-   *
-   * @param spark  spark session
-   * @param config [[com.typesafe.config.Config]] typesafe Config object
-   */
-  def this(spark: SparkSession, config: Config) = this(config)
 
   private[this] val cqlConnection =  DatastaxCassandraConnector(spark.sparkContext.getConf)