+++ src/main/scala/com/jcdecaux/setl/storage/connector/ExcelConnector.scala
@@ -3,6 +3,7 @@ package com.jcdecaux.setl.storage.connector
 import com.jcdecaux.setl.annotation.InterfaceStability
 import com.jcdecaux.setl.config.Conf
 import com.jcdecaux.setl.enums.Storage
+import com.jcdecaux.setl.internal.HasReaderWriter
 import com.jcdecaux.setl.util.TypesafeConfigUtils
 import com.typesafe.config.Config
 import org.apache.spark.sql._
@@ -40,7 +41,7 @@ class ExcelConnector(val path: String,
                      var workbookPassword: Option[String],
                      var schema: Option[StructType],
                      var saveMode: SaveMode
-                    ) extends Connector {
+                    ) extends Connector with HasReaderWriter {
 
   // CONSTRUCTORS
   def this(spark: SparkSession,
@@ -120,11 +121,6 @@ class ExcelConnector(val path: String,
     saveMode = SaveMode.valueOf(TypesafeConfigUtils.getAs[String](config, "saveMode").getOrElse("Overwrite"))
   )
 
-  @deprecated("use the constructor with no spark session", "0.3.4")
-  def this(spark: SparkSession, conf: Conf) = this(conf)
-
-  @deprecated("use the constructor with no spark session", "0.3.4")
-  def this(spark: SparkSession, config: Config) = this(config)
   //END CONSTRUCTOR
 
   if (sheetName.isDefined) {